requirements.txt
requests~=2.28.1
pandas~=1.5.2
jsonpath~=0.82.2
numpy~=1.21.5
matplotlib~=3.6.2
scikit-learn~=1.2.0
jieba~=0.42.1
gensim~=4.3.2
nltk~=3.7
wordcloud~=1.9.2
pillow~=9.3.0
pyldavis~=3.4.0
adjusttext~=0.8